Find any most frequent Maxigram.
In this program you are given a string s consisting of n capital Latin letters. Your task is to find any Maxigram contained in the given string as a substring (i.e. two consecutive characters of the string) maximal number of times. For example, for string s= "BBAABBBA" the answer is Maxigram "BB", which contained in s three times. In other words, find any most frequent Maxigram.